+++ tests/integration/test_get_species_list.py
@@ -1,0 +1,6 @@
+from ebird.api import get_species_list
+
+
+def test_api_call(api_token, subnational1_code):
+    species = get_species_list(api_token, subnational1_code)
+    assert len(species)